pyproject.toml
[tool.poetry]
name = "watts-insight"
version = "1.2"
description = ""
authors = ["Simon Kammerer <simon.kammerer@volterica.com>"]
readme = "README.md"
packages = [{include = "factory_flexibility_model"}]

[tool.poetry.dependencies]
python = ">=3.11, <4"
dash = "^2.11.1"
dash-auth = "^1.4.1"
dash-bootstrap-components = "^1.4.2"
dash-bootstrap-templates = "^1.0.8"
datetime = "^5.2"
gurobipy = "^11.0.0"
joblib = "^1.3.1"
kivy = "^2.2.1"
kivymd = "^1.2.0"
kivy-garden-graph = "^0.4.0"
matplotlib = "^3.7.2"
numpy = "^1.25.2"
openpyxl = "^3.1.2"
pandas = "^2.0.3"
plotly = "^5.15.0"
scipy = "^1.11.1"
xlsxwriter = "^3.1.2"
pyyaml = "^6.0.1"
python-decouple = "^3.8"
sphinx = "^7.2.6"
sphinx-rtd-theme = "^1.3.0"
optuna = "^3.6.1"
optuna-dashboard = "^0.15.1"
psycopg2 = "^2.9.9"
botorch = "^0.11.1"
optuna-integration = "^3.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gui = "examples.main:gui"
dashboard = "examples.main:dash"
simulate = "examples.simple_simulation_call:simulate"
optimize = "examples.main:run_optimizer"
simulate_session = "examples.main:simulate_session"
create_shortcut = "examples.create_desktop_shortcut:create_desktop_shortcut"
